pyproject.toml
[project]
name = "pdf-extract-pipeline"
version = "0.1.0"
description = "Schema-driven PDF extraction pipeline powered by Gemini. Turn messy PDFs into clean structured data."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "joshobuilds" }]
dependencies = [
    "google-genai>=0.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
pdf-extract = "pdf_extract.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_extract"]
